+++ Connecto.DataObjects/EntityFramework/EntityProductDetail.cs
@@ -17,7 +17,7 @@ namespace Connecto.DataObjects.EntityFramework
         public double UnitPrice { get; set; }
         public double SellingPrice { get; set; }
         public int Quantity { get; set; }
-        public int QuantityLower { get; set; }
+        public double QuantityLower { get; set; }
         public DateTime DateReceived { get; set; }
         public virtual EntityOrder Order { get; set; }
         public virtual EntityProduct Product { get; set; }
